adicionando funções

+++ pages/cursos/index.js
@@ -4,7 +4,7 @@ import React, { useEffect, useState } from 'react'
 import { Table } from 'react-bootstrap'
 import { AiOutlinePlus } from 'react-icons/ai'
 import { BsTrash3 } from 'react-icons/bs'
-
+import { HiOutlinePencilAlt } from 'react-icons/hi'
 
 const index = () => {
 
@@ -14,22 +14,24 @@ const index = () => {
         setCursos(getAll())
     }, [])
 
-    function getAll(){
+    function getAll() {
         return JSON.parse(window.localStorage.getItem('cursos')) || []
     }
 
-    function excluir(id){
-        const itens = getAll()
-        itens.splice(id, 1)
-        window.localStorage.setItem('cursos', JSON.stringify(itens))
-        setCursos(itens)
+    function excluir(id) {
+        if (confirm('Deseja realmente excluir o registro? ')) {
+            const itens = getAll()
+            itens.splice(id, 1)
+            window.localStorage.setItem('cursos', JSON.stringify(itens))
+            setCursos(itens)
+        }
     }
 
     return (
         <Pagina titulo='Cursos'>
 
             <Link href="/cursos/form" className='mb-2 btn btn-primary'>
-                <AiOutlinePlus />
+                <AiOutlinePlus className='me-1' />
                 Novo
             </Link>
             <Table striped bordered hover >
@@ -46,7 +48,11 @@ const index = () => {
                     {cursos.map((item, i) => (
                         <tr key={i}>
                             <td>
-                                <BsTrash3 onClick={() => excluir(i)} className='text-warning' />
+                                <Link href={'/cursos/' + 1}>
+                                    <HiOutlinePencilAlt title='Alterar' className='text-primary' />
+                                </Link>
+                                {' '}
+                                <BsTrash3 title='Excluir' onClick={() => excluir(i)} className='text-warning' />
                             </td>
                             <td>{item.nome}</td>
                             <td>{item.duracao}</td>
